requirements.txt
fastapi
uvicorn
openai
pdfplumber
python-docx
pandas
python-dotenv